move summoned info to playerdata and don't show 0 at summoned count
