+++ src/main/java/com/robertx22/mine_and_slash/capability/entity/EntityData.java
@@ -137,7 +137,6 @@ public class EntityData implements ICap, INeededForClient {
     private static final String LEECH = "leech";
     private static final String MAP_ID = "mapid";
     private static final String MAP_MOB = "map_mob";
-    private static final String SUMMONED = "summoned";
 
     private transient int dontSyncTicks = 0;
 
@@ -194,7 +193,6 @@ public class EntityData implements ICap, INeededForClient {
     CooldownsData cooldowns = new CooldownsData();
     ThreatData threat = new ThreatData();
     public EntityLeechData leech = new EntityLeechData();
-    SummonedData summonedData = new SummonedData();
 
     public boolean didStatCalcThisTickForPlayer = false;
 
@@ -286,7 +284,6 @@ public class EntityData implements ICap, INeededForClient {
             LoadSave.Save(summonedPetData, nbt, PET);
             LoadSave.Save(leech, nbt, LEECH);
             LoadSave.Save(customExactStats, nbt, CUSTOM_STATS);
-            LoadSave.Save(summonedData, nbt, SUMMONED);
 
 
             if (customExactStats != null) {
@@ -352,7 +349,6 @@ public class EntityData implements ICap, INeededForClient {
             this.resources = loadOrBlank(ResourcesData.class, new ResourcesData(), nbt, RESOURCES_LOC, new ResourcesData());
             this.cooldowns = loadOrBlank(CooldownsData.class, new CooldownsData(), nbt, COOLDOWNS, new CooldownsData());
             this.leech = loadOrBlank(EntityLeechData.class, new EntityLeechData(), nbt, LEECH, new EntityLeechData());
-            this.summonedData = loadOrBlank(SummonedData.class, new SummonedData(), nbt, SUMMONED, new SummonedData());
 
 
         } catch (Exception e) {
@@ -439,10 +435,6 @@ public class EntityData implements ICap, INeededForClient {
 
     public ThreatData getThreat() {
         return threat;
-    }
-
-    public SummonedData getSummonedData() {
-        return summonedData;
     }
 
     public Unit getUnit() {
@@ -984,10 +976,5 @@ public class EntityData implements ICap, INeededForClient {
     @Override
     public String getCapIdForSyncing() {
         return "entity_data";
-    }
-
-    public void addSummonedType(String spellId, int amount) {
-        this.summonedData.addSummonedType(spellId, amount);
-        this.sync.setDirty();
     }
 }
