+++ src/main/java/com/robertx22/mine_and_slash/gui/overlays/spell_hotbar/SpellOnHotbarRender.java
@@ -108,8 +108,8 @@ public class SpellOnHotbarRender {
 
                 }
 
-                SummonedData summonedData = Load.Unit(mc.player).getSummonedData();
-                if (summonedData.hasSummoned(spell.GUID())) {
+                SummonedData summonedData = Load.player(mc.player).getSummonedData();
+                if (summonedData.getSummonedAmount(spell.GUID()) > 0) {
                     drawSummoned(xs + 15, ys + 3, summonedData.getSummonedAmount(spell.GUID()));
                 }
